// src/notifications/enums/notification-type.enum.ts
export enum NotificationType {
  LEAVE_REQUEST = 'LEAVE_REQUEST',
  LEAVE_APPROVED = 'LEAVE_APPROVED',
  LEAVE_REJECTED = 'LEAVE_REJECTED',
  PERFORMANCE_REVIEW = 'PERFORMANCE_REVIEW',
  ATTENDANCE = 'ATTENDANCE',
  EMPLOYEE_UPDATE = 'EMPLOYEE_UPDATE',
  SYSTEM = 'SYSTEM',
  TASK_ASSIGNED = 'TASK_ASSIGNED',
  TASK_UPDATED = 'TASK_UPDATED',
  PAYROLL = 'PAYROLL',
}